requirements.txt
backcall==0.2.0
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
feature-engine==0.5.15
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.16.0
kiwisolver==1.2.0
lark-parser==0.9.0
matplotlib==3.3.0
numpy==1.19.1
pandas==1.1.0
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
statsmodels==0.11.1
threadpoolctl==2.1.0
tqdm==4.48.2
traitlets==4.3.3
wcwidth==0.2.5
z3-solver==4.8.8.0
